toasty_sql::stmt

Enum ExprSet

pub enum ExprSet {
    Select(Box<Select>),
    SetOp(ExprSetOp),
    Update(Box<Update>),
    Values(Values),
    Insert(Box<Insert>),
}

A set of rows produced by a query, set operation, or explicit values.

Represents the different ways to produce a collection of rows in SQL.

Examples

SELECT * FROM users           // ExprSet::Select
SELECT ... UNION SELECT ...   // ExprSet::SetOp
VALUES (1, 'a'), (2, 'b')     // ExprSet::Values

Variants

Select(Box<Select>)

A select query, possibly with a filter.

SetOp(ExprSetOp)

A set operation (union, intersection, …) on two queries.

Update(Box<Update>)

An update expression.

Values(Values)

Explicitly listed values (as expressions).

Insert(Box<Insert>)

An insert statement (used for UNION-style batch inserts)

Implementations

impl ExprSet

pub fn values(values: impl Into<Values>) -> ExprSet

Creates an ExprSet::Values from explicit values.

pub fn is_values(&self) -> bool

Returns true if this is an ExprSet::Values variant.

Examples

let values = ExprSet::values(Values::default());
assert!(values.is_values());

let select = ExprSet::from(toasty_core::schema::db::TableId(0));
assert!(!select.is_values());

pub fn as_values(&self) -> Option<&Values>

Returns a reference to the inner Values if this is an ExprSet::Values.

Returns None for all other ExprSet variants.

pub fn as_values_unwrap(&self) -> &Values

Returns a reference to the inner Values.

Panics

Panics if self is not an ExprSet::Values.

pub fn as_values_mut(&mut self) -> Option<&mut Values>

Returns a mutable reference to the inner Values if this is an ExprSet::Values, or None otherwise.

pub fn as_values_mut_unwrap(&mut self) -> &mut Values

Returns a mutable reference to the inner Values.

Panics

Panics if self is not an ExprSet::Values.

pub fn into_values(self) -> Values

Consumes the expression set and returns the inner Values.

Panics

Panics (via todo!()) if self is not an ExprSet::Values.

pub fn is_const(&self) -> bool

Returns true if this expression set contains only constant expressions (no references, subqueries, or other external data).

impl ExprSet

pub fn as_select(&self) -> Option<&Select>

Returns a reference to the inner Select if this is a Select variant.

pub fn as_select_unwrap(&self) -> &Select

Returns a reference to the inner Select, panicking if this is not a Select variant.

pub fn as_select_mut(&mut self) -> Option<&mut Select>

Returns a mutable reference to the inner Select if this is a Select variant.

pub fn as_select_mut_unwrap(&mut self) -> &mut Select

Returns a mutable reference to the inner Select, panicking if this is not a Select variant.

pub fn into_select(self) -> Select

Consumes this ExprSet and returns the inner Select.

Panics

Panics if this is not a Select variant.

pub fn is_select(&self) -> bool

Returns true if this is a Select variant.